requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.6.1
jwt==1.3.1
bcrypt==4.1.2